backend/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
python-dotenv~=0.21.0
gunicorn
whitenoise
ipython

